Create Chris Spangler's FinalProjectBRFSS.html
This project uses the behavioral risk factor surveillance system data collected in 2013. Using the data, appropriate research questions, and an exploratory data analysis, correlations are drawn between 1: Obesity and high cholesterol, 2: Obesity and cholesterol affecting men and women similarly or differently, and 3: An examination of health care coverage for people who are obese and have high levels of cholesterol. The project uses both categorical variables and appropriate visualizations. The variables are not weighted, yet they show a correlation between being overweight/obese and having a high cholesterol level. Males are twice as likely to have weight issues, according to the data. The vast majority of people who have medical conditions do in fact have health insurance. This is a correlative study, but not a causal analysis.